Rework browse-on-ssh-client
Instead of trying to directly invoke a browser function, which has
trouble on Windows due to logon type interactions, and is generally
unreliable, just drop a file in a designated directory.  Set up a
polling daemon on the remote machine to monitor that directory and
open the URL in each file.